refactor(common): move source version validation logic to SourceChildMixin
* Move belongs_to_non_head_source_version to SourceChildMixin for reuse across models

* Move has_pending_source_version_seed to SourceChildMixin to centralize validation logic
